pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "Betelgeusse",email = "betelgeussesoftware@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai (>=1.101.0,<2.0.0)",
    "langchain-openai (>=0.3.31,<0.4.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "fastapi[standard] (>=0.116.1,<0.117.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
